Rename `Future::select2` to `select`
... and remove the previous `select`
